rename: 이벤트 생성 이미지 파일 생성 폴더로 분리
