+++ src/features/event-manage/ui/EventCategory.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import CategoryButton from '../../../../public/assets/event-manage/BackBtn(black).svg';
+import CategoryButton from '../../../../public/assets/event-manage/creation/BackBtn(black).svg';
 import { FunnelState } from '../model/FunnelContext';
 
 interface Category {
