+++ src/features/event-manage/ui/EventType.tsx
@@ -1,5 +1,5 @@
-import OnlineIcon from '../../../../public/assets/event-manage/OnlineIcon.svg';
-import OfflineIcon from '../../../../public/assets/event-manage/OfflineIcon.svg';
+import OnlineIcon from '../../../../public/assets/event-manage/creation/OnlineIcon.svg';
+import OfflineIcon from '../../../../public/assets/event-manage/creation/OfflineIcon.svg';
 import SearchBar from '../../../shared/ui/SearchBar';
 import { useFunnelState } from '../model/FunnelContext';
