+++ src/features/event-manage/ui/FileUpload.tsx
@@ -1,4 +1,4 @@
-import FileUploadImage from '../../../../public/assets/event-manage/FileUpload.svg';
+import FileUploadImage from '../../../../public/assets/event-manage/creation/FileUpload.svg';
 
 const FileUpload = () => {
   return (
